Cap nhat thong tin tai khoan, dang nhap, dang xuat

+++ user/php/XLthanhtoan.php
@@ -6,25 +6,24 @@ $conn = connectDB();
 if (!isset($_POST['tonggiatri']) || !isset($_POST['maGG']) || !isset($_POST['maVC']) || !isset($_POST['maKH']) || !isset($_POST['ngay'])) {
     echo "aaaa";
 }
-$tongGiaTri = $_POST['tonggiatri'];
-$maGG = $_POST['maGG'];
-$maVC = $_POST['maVC'];
-$maKH = $_POST['maKH'];
-$ngay = $_POST['ngay'];
-$trangthai = 0;
-$stmt = mysqli_prepare($conn, "INSERT INTO donhang(Trangthai, Ngay, Tonggiatri, Magiamgia, Mavc, maKhachhang) VALUE (?,?,?,?,?,?)");
-mysqli_stmt_bind_param($stmt, 'isssss', $trangthai, $ngay, $tongGiaTri, $maGG, $maVC, $maKH);
-$rs = mysqli_stmt_execute($stmt);
-if ($rs === false) {
-    die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
-} else {
-    // Nếu không có lỗi, in ra thông báo cho biết INSERT đã thành công
-    $maDH = mysqli_insert_id($conn);
-    $sqlGioHang = "SELECT * FROM giohang WHERE Manguoidung='$maKH'";
-    $rsGioHang = mysqli_query($conn, $sqlGioHang);
-    while ($row = mysqli_fetch_array($rsGioHang)) {
-        $maSP = $row['Masp'];
-        $soLuong = $row['Soluong'];
+if (isset($_POST['maSP']) && isset($_POST['soLuong'])) {
+    $tongGiaTri = $_POST['tonggiatri'];
+    $maGG = $_POST['maGG'];
+    $maVC = $_POST['maVC'];
+    $maKH = $_POST['maKH'];
+    $ngay = $_POST['ngay'];
+    $trangthai = 0;
+    $maSP = $_POST['maSP'];
+    $soLuong = $_POST['soLuong'];
+    $stmt = mysqli_prepare($conn, "INSERT INTO donhang(Trangthai, Ngay, Tonggiatri, Magiamgia, Mavc, maKhachhang) VALUE (?,?,?,?,?,?)");
+    mysqli_stmt_bind_param($stmt, 'isssss', $trangthai, $ngay, $tongGiaTri, $maGG, $maVC, $maKH);
+    $rs = mysqli_stmt_execute($stmt);
+    if ($rs === false) {
+        die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+    } else {
+        // Nếu không có lỗi, in ra thông báo cho biết INSERT đã thành công
+        $maDH = mysqli_insert_id($conn);
+
         $sqlCTDH = "INSERT INTO chitietdonhang(Masp, Madonhang, Soluong) VALUE ($maSP, $maDH, $soLuong)";
         $rsCTDH = mysqli_query($conn, $sqlCTDH);
         if ($rsCTDH === false) {
@@ -32,14 +31,64 @@ if ($rs === false) {
         } else {
             echo "Insert thành công CTDH!";
         }
-    }
-    $delete = "DELETE FROM giohang WHERE Manguoidung='$maKH'";
-    $rsDelete = mysqli_query($conn, $delete);
-    if ($rsDelete) {
-        echo 'Xóa thành công giỏ hàng của ng dùng hiện tại';
-    } else {
-        die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
-    }
+        $sqlCapNhatSoLuong = "UPDATE sanpham SET Soluongconlai=Soluongconlai-? WHERE Masp=?";
+        $stmtCNSL = mysqli_prepare($conn, $sqlCapNhatSoLuong);
+        mysqli_stmt_bind_param($stmtCNSL, 'is', $soLuong, $maSP);
+        $rsCNSL = mysqli_stmt_execute($stmtCNSL);
+        if ($rsCNSL === false) {
+            die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+        } else {
+            echo "Cập nhật số lượng thành công!";
+        }
 
-    echo "Insert thành công!";
+
+        echo "Insert thành công!";
+    }
+} else {
+    $tongGiaTri = $_POST['tonggiatri'];
+    $maGG = $_POST['maGG'];
+    $maVC = $_POST['maVC'];
+    $maKH = $_POST['maKH'];
+    $ngay = $_POST['ngay'];
+    $trangthai = 0;
+    $stmt = mysqli_prepare($conn, "INSERT INTO donhang(Trangthai, Ngay, Tonggiatri, Magiamgia, Mavc, maKhachhang) VALUE (?,?,?,?,?,?)");
+    mysqli_stmt_bind_param($stmt, 'isssss', $trangthai, $ngay, $tongGiaTri, $maGG, $maVC, $maKH);
+    $rs = mysqli_stmt_execute($stmt);
+    if ($rs === false) {
+        die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+    } else {
+        // Nếu không có lỗi, in ra thông báo cho biết INSERT đã thành công
+        $maDH = mysqli_insert_id($conn);
+        $sqlGioHang = "SELECT * FROM giohang WHERE Manguoidung='$maKH'";
+        $rsGioHang = mysqli_query($conn, $sqlGioHang);
+        while ($row = mysqli_fetch_array($rsGioHang)) {
+            $maSP = $row['Masp'];
+            $soLuong = $row['Soluong'];
+            $sqlCTDH = "INSERT INTO chitietdonhang(Masp, Madonhang, Soluong) VALUE ($maSP, $maDH, $soLuong)";
+            $rsCTDH = mysqli_query($conn, $sqlCTDH);
+            if ($rsCTDH === false) {
+                die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+            } else {
+                echo "Insert thành công CTDH!";
+            }
+            $sqlCapNhatSoLuong = "UPDATE sanpham SET Soluongconlai=Soluongconlai-? WHERE Masp=?";
+            $stmtCNSL = mysqli_prepare($conn, $sqlCapNhatSoLuong);
+            mysqli_stmt_bind_param($stmtCNSL, 'is', $soLuong, $maSP);
+            $rsCNSL = mysqli_stmt_execute($stmtCNSL);
+            if ($rsCNSL === false) {
+                die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+            } else {
+                echo "Cập nhật số lượng thành công!";
+            }
+        }
+        $delete = "DELETE FROM giohang WHERE Manguoidung='$maKH'";
+        $rsDelete = mysqli_query($conn, $delete);
+        if ($rsDelete) {
+            echo 'Xóa thành công giỏ hàng của ng dùng hiện tại';
+        } else {
+            die("Lỗi trong quá trình thực hiện câu lệnh: " . mysqli_error($conn));
+        }
+
+        echo "Insert thành công!";
+    }
 }
